power: Introduce Quick Charge v1.0
First of all, I would like to thank Varun Chitre for Developing a Great Fast-Charge Driver i.e., ThunderCharge for Tomato(+) Devices. Before this Driver, the Charging Time of Tomato(+) Devices was very long, sometimes even exceeding 2 and a Half-Hours! However, with the Introduction of ThunderCharge, this Slow-Charging issue was Solved. So, Thank You Varun Chitre for your Wonderful Contribution.

Now, in ThunderCharge, I observed a Potential to Achieve even More Efficiency while Charging. In my opinion, ThunderCharge suffers from the following Problems---
1. A Constant Current (mA) is Forced at all the Times of Charging without any Attention to Battery Status.
2. The above Method, I think, proves Quite Dangerous if Higher Values of Current (mA) are used because the Current (mA) remains Constant throughout the Charge-Cycle.
3. All of this creates a rigidness in the Driver and thus, Higher Values of Current (mA) are rarely used because of Probable Damage to Hardware.
4. Since all the Tomato(+) Devices suffer from Over-Heating issues, Forcing a Constant Current (mA) at all the Times of Charging may Prove Very Dangerous.

To Fix these issues, I decided to Alter the Mechanism of ThunderCharge Current Control Driver. As you all know, Qualcomm's Quick-Charge Technology Charges the Phones Quite Fast. In order to achieve this, the Battery is first Charged upto a certain % with the Max. Supported Current (mA), then the Value of Current (mA) is Dropped Gradually as the Battery nears 100%. As per the Explaination of this Technology, this is done to Avoid Damage to the Battery and Keep the Temperatures in Check. Apparently, Charging the Battery at a Constant High Current (mA) when the Battery is about to get Fully Charged, is very Dangerous and Degrades the Battery-Life quickly.

On the Basis of such Information and Technique, I have Altered the Working of ThunderCharge and turned this into a New Driver called Quick Charge v1.0. Moreover, I have Removed most of the sysFS Controls too as they are no longer Required.

Explaination of the Working of Quick Charge v1.0, developed by Shoaib Anwar a.k.a. Shoaib0597---
Battery (%)				Current (mA)
0-60					1500 (Max. Supported as per DTB)
61-90					1250
91-100					1000 (Standard Output)

Notes---
-This is NOT the Original Quick Charge v1.0 Technology from Qualcomm. I have just used the Name because the Working/Mechanism is Similar.
-The Stock Hardware doesn't supports Qualcomm's Quick Charge v1.0 Technology.
-Thus, the Charging Speed won't be as Fast as Qualcomm's Quick Charge v1.0.
-However, the Charging Speed will be the Fastest Possible on the Stock Hardware.
-In other words, this Driver is an Improved Version of ThunderCharge and Offers the Best Possible Experience and Efficiency.

Signed-off-by: Shoaib0597 <Shoaib0595@gmail.com>
Signed-off-by: nancyjudge <botnancy1@gmail.com>